# adodasopenapi.yaml
openapi: 3.0.3
info:
  title: Adodas API
  description: |
    API REST completa para gestión de **clientes, ítems, pedidos, reembolsos y servicios**.
    
    ## ✨ Características:
    - ✅ **CRUD completo** todos los recursos
    - ✅ **Validaciones robustas** (DNI, emails, precios)
    - ✅ **Filtros avanzados** por estado, precio, cliente
    - ✅ **Errores específicos** 400/404/500 diferenciados
    - ✅ **Ejemplos reales** en todas las respuestas
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Desarrollo Local
...
components:
  schemas:
    ErrorResponseDto:
      type: object
      properties:
        code:
          type: integer
          description: Código HTTP del error
        message:
          type: string
          description: Descripción del error
      required: [ code, message ]
    CustomerDto:
      type: object
      properties:
        name:
          type: string
          example: "Juan"
        surname:
          type: string
          example: "Pérez"
        dni:
          type: string
          pattern: '^\d{8}[A-Z]$'
          example: "12345678A"
        email:
          type: string
          format: email
          example: "juan.perez@email.com"
        isEmailConfirmed:
          type: boolean
          example: true
        registerDate:
          type: string
          format: date
          readOnly: true
          example: "2026-01-15"
...
    CustomerV2Dto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 10
        name:
          type: string
          description: Nombre del cliente.
          example: Juan
        email:
          type: string
          format: email
          description: Correo electrónico del cliente.
          example: juan.perez@example.com
        phone:
          type: string
          description: Teléfono del cliente.
          example: "+34 600 123 456"
      required:
        - name
        - email
        - phone
paths:
  /v1/customers:
    get:
      tags: [ CUSTOMERS ]
      summary: Lista todos los clientes
      description: Obtiene una lista paginada o filtrada de clientes registrados.
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Filtra por nombre del cliente.
        - name: email
          in: query
          required: false
          schema:
            type: string
          description: Filtra por email del cliente.
        - name: surname
          in: query
          required: false
          schema:
            type: string
          description: Filtra por apellido del cliente.
      responses:
        '200':
          description: Lista de clientes obtenida correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerDto'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    post:
      tags: [ CUSTOMERS ]
      summary: Crea un nuevo cliente
      description: Registra un nuevo cliente validando DNI, email y formato de datos.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerDto'
      responses:
        '201':
          description: Cliente creado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDto'
        '400':
          description: Error de validación en los datos proporcionados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 400
                message: "Validation failed"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"
  /v1/customers/{customerId}:
    get:
      tags: [ CUSTOMERS ]
      summary: Obtiene un cliente por su ID
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del cliente.
      responses:
        '200':
          description: Cliente encontrado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDto'
        '404':
          description: Cliente no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Cliente no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    put:
      tags: [ CUSTOMERS ]
      summary: Modifica un cliente existente
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del cliente a modificar.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerDto'
      responses:
        '200':
          description: Cliente modificado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDto'
        '400':
          description: Error de validación en los datos proporcionados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 400
                message: "Validation failed"
        '404':
          description: Cliente no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Cliente no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    patch:
      tags: [ CUSTOMERS ]
      summary: Actualiza parcialmente un cliente
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del cliente a actualizar parcialmente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                email: "nuevo.email@email.com"
      responses:
        '200':
          description: Cliente actualizado parcialmente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDto'
        '400':
          description: Error de validación o cuerpo inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 400
                message: "Validation failed"
        '404':
          description: Cliente no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Cliente no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    delete:
      tags: [ CUSTOMERS ]
      summary: Elimina un cliente
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del cliente a eliminar.
      responses:
        '204':
          description: Cliente eliminado correctamente (sin contenido en la respuesta).
        '404':
          description: Cliente no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Cliente no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"
  /v2/customers:
    get:
      tags: [ CUSTOMERS ]
      summary: Lista todos los clientes v2
      description: Obtiene una lista paginada de clientes usando page y size.
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Número de página, empezando en 0.
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 5
          description: Tamaño de página.
      responses:
        '200':
          description: Lista de clientes obtenida correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerDto'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    post:
      tags: [ CUSTOMERS ]
      summary: Crea un nuevo cliente v2
      description: Registra un cliente usando el contrato v2 con phone obligatorio.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerV2Dto'
      responses:
        '201':
          description: Cliente creado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerV2Dto'
        '400':
          description: Error de validación en los datos proporcionados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 400
                message: "Validation failed"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"
  /v2/customers/{id}:
    put:
      tags: [ CUSTOMERS ]
      summary: Actualiza un cliente v2
      description: Solo permite modificar email y surname.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del cliente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                email: nuevo@email.com
                surname: García
      responses:
        '200':
          description: Cliente actualizado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDto'
        '404':
          description: Cliente no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Cliente no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    delete:
      tags: [ CUSTOMERS ]
      summary: Elimina un cliente v2
      description: Simula un soft delete y devuelve un mensaje de confirmación.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del cliente.
      responses:
        '200':
          description: Cliente marcado como eliminado correctamente.
          content:
            text/plain:
              schema:
                type: string
              example: Customer with id 10 marked as deleted (soft delete in v2)
        '404':
          description: Cliente no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Cliente no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"
...